+++ Pods/Wit/Wit/WITContextSetter.m
@@ -9,25 +9,8 @@
 #import "WITContextSetter.h"
 #import "WitPrivate.h"
 #import "util.h"
-#import <CoreLocation/CoreLocation.h>
 
-@implementation WITContextSetter {
-    CLLocationManager *locationManager;
-}
-
-
-
--(void)ensureLocation:(NSMutableDictionary *)context {
-    if ([self locationAccess] == NO) {
-        return ;
-    }
-    CLLocationDegrees latitude = locationManager.location.coordinate.latitude;
-    CLLocationDegrees longitude = locationManager.location.coordinate.longitude;
-    NSNumber *oLatitude = [[NSNumber alloc] initWithDouble:latitude];
-    NSNumber *oLongitude = [[NSNumber alloc] initWithDouble:longitude];
-    NSDictionary *locationData = [[NSDictionary alloc] initWithObjectsAndKeys:oLatitude, @"latitude", oLongitude, @"longitude", nil];
-    [context setObject:locationData forKey:@"location"];
-}
+@implementation WITContextSetter
 
 -(void)ensureReferenceTime:(NSMutableDictionary *)context {
     NSDateFormatter *dateFormatter = [[NSDateFormatter alloc] init];
@@ -41,41 +24,12 @@
     
 }
 
--(BOOL)locationAccess {
-    if (locationManager == nil) {
-        locationManager = [[CLLocationManager alloc] init];
-    }
-    
-    
-    CLAuthorizationStatus currentStatus = [CLLocationManager authorizationStatus];
-    if (currentStatus == kCLAuthorizationStatusDenied
-        || currentStatus == kCLAuthorizationStatusRestricted) {
-        return NO;
-    }
-    if (currentStatus == kCLAuthorizationStatusNotDetermined
-        && [locationManager respondsToSelector:@selector(requestWhenInUseAuthorization)]) {
-        [locationManager requestWhenInUseAuthorization];
-        currentStatus = [CLLocationManager authorizationStatus];
-    }
-    if (currentStatus == kCLAuthorizationStatusDenied
-        || currentStatus == kCLAuthorizationStatusRestricted) {
-        return NO;
-    }
-    NSLog(@"Location access status: %d", currentStatus);
-    [locationManager startMonitoringSignificantLocationChanges];
-    
-    return YES;
-}
-
 -(void)contextFillup:(NSMutableDictionary *)context {
-    [self ensureLocation:context];
     [self ensureReferenceTime:context];
 }
 
 -(id)init {
     self = [super init];
-    [self locationAccess];
-    
     return self;
 }